# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API documentation",
        "title": "Project Management App",
        "contact": {
            "email": "m.andres.novrizal@gmail.com"
        }
    },
    "paths": {
...
        "/board": {
            "post": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Create a new board. This endpoint requires cookie authentication.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "boards"
                ],
                "summary": "Create a new board",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Name of the board",
                        "name": "name_board",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/web.BoardResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/web.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/web.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/web.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/board/{id}": {
            "put": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Edit a board's name by its ID.  This endpoint requires cookie authentication.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "boards"
                ],
                "summary": "Edit a board",
                "parameters": [
                    {
                        "minimum": 1,
                        "type": "integer",
                        "example": 1,
                        "description": "Board ID parameter",
                        "name": "request",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "New name of the board",
                        "name": "name_board",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/web.BoardResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/web.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/web.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/web.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Delete a board by its ID.  This endpoint requires cookie authentication.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "boards"
                ],
                "summary": "Delete a board",
                "parameters": [
                    {
                        "minimum": 1,
                        "type": "integer",
                        "example": 1,
                        "description": "Board ID parameter",
                        "name": "request",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/web.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/web.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/web.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/web.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/boards": {
            "get": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Retrieve all boards.  This endpoint requires cookie authentication.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "boards"
                ],
                "summary": "Get all boards",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/web.BoardResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/web.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/web.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/web.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "web.BoardResponse": {
            "type": "object",
            "properties": {
                "board_created_by": {
                    "type": "object",
                    "properties": {
                        "user_email": {
                            "type": "string",
                            "example": "user@example.com"
                        },
                        "user_id": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                },
                "board_id": {
                    "type": "integer",
                    "example": 1
                },
                "name_board": {
                    "type": "string",
                    "example": "My Project Board"
                },
                "tasks": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/web.TaskInfo"
                    }
                }
            }
        },
...
        "web.EmployeeWithInvitation": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "example": "employee@example.com"
                },
                "invitation_status": {
                    "type": "string",
                    "example": "Pending"
                },
                "user_id": {
                    "type": "integer",
                    "example": 3
                }
            }
        },
        "web.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Error Message"
                }
            }
        },
...
        "web.ManagerWithInvitation": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "example": "manager@example.com"
                },
                "invitation_status": {
                    "type": "string",
                    "example": "Accepted"
                },
                "user_id": {
                    "type": "integer",
                    "example": 2
                }
            }
        },
        "web.Owner": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "example": "owner@example.com"
                },
                "user_id": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "web.PlanningDescriptionFile": {
            "type": "object",
            "properties": {
                "file_name": {
                    "type": "string",
                    "example": "planning_description.pdf"
                },
                "file_url": {
                    "type": "string",
                    "example": "https://bucket-name.s3.amazonaws.com/planning_description.pdf"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "web.PlanningFile": {
            "type": "object",
            "properties": {
                "file_name": {
                    "type": "string",
                    "example": "planning_document.docx"
                },
                "file_url": {
                    "type": "string",
                    "example": "https://bucket-name.s3.amazonaws.com/planning_document.docx"
                },
                "id": {
                    "type": "integer",
                    "example": 2
                }
            }
        },
        "web.ProjectFile": {
            "type": "object",
            "properties": {
                "file_name": {
                    "type": "string",
                    "example": "project_report.pdf"
                },
                "file_url": {
                    "type": "string",
                    "example": "https://bucket-name.s3.amazonaws.com/project_report.pdf"
                },
                "id": {
                    "type": "integer",
                    "example": 3
                }
            }
        },
...
        "web.SuccessResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Success Message"
                }
            }
        },
...
        "web.TaskInfo": {
            "type": "object",
            "properties": {
                "board_id": {
                    "type": "integer",
                    "example": 1
                },
                "employee": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/web.EmployeeWithInvitation"
                    }
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "manager": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/web.ManagerWithInvitation"
                    }
                },
                "name_task": {
                    "type": "string",
                    "example": "Implement feature X"
                },
                "owner": {
                    "$ref": "#/definitions/web.Owner"
                },
                "planning_description_files": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/web.PlanningDescriptionFile"
                    }
                },
                "planning_description_persen": {
                    "type": "string",
                    "example": "50"
                },
                "planning_due_date": {
                    "type": "string",
                    "example": "2023-12-31"
                },
                "planning_file": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/web.PlanningFile"
                    }
                },
                "planning_status": {
                    "type": "string",
                    "example": "Approved"
                },
                "priority": {
                    "type": "string",
                    "example": "High"
                },
                "project_comment": {
                    "type": "string",
                    "example": "Making good progress"
                },
                "project_due_date": {
                    "type": "string",
                    "example": "2024-01-15"
                },
                "project_file": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/web.ProjectFile"
                    }
                },
                "project_status": {
                    "type": "string",
                    "example": "Undone"
                }
            }
        },
...
    }
}